Formatting

+++ PoolMemberships/Data/PoolMembershipDbContext.cs
@@ -13,21 +13,14 @@ public class PoolMembershipDbContext : DbContext
 
     protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
     {
-        string dbPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "PMM");
+        var dbPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "PMM");
 
-        if (!Directory.Exists(dbPath))
-        {
-            Directory.CreateDirectory(dbPath);
-        }
+        if (!Directory.Exists(dbPath)) Directory.CreateDirectory(dbPath);
 
-        string dbFile = Path.Combine(dbPath, "PoolMembershipDatabase.db");
+        var dbFile = Path.Combine(dbPath, "PoolMembershipDatabase.db");
 
-        if (!File.Exists(dbFile))
-        {
-            File.Create(dbFile).Close();
-        }
-        
+        if (!File.Exists(dbFile)) File.Create(dbFile).Close();
+
         optionsBuilder.UseSqlite($"Data Source={dbFile}");
-    } 
-    
+    }
 }
